refactor: move push logic out of create entry service
